backend-ml/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
scikit-image>=0.20.0
opencv-python-headless>=4.7.0
scipy>=1.10.0
psycopg2-binary>=2.9.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
